move `CollateralPower` to `dango_types::account::margin`
