+++ dango/types/src/account/margin.rs
@@ -1,4 +1,16 @@
-use grug::Udec128;
+use grug::{Bound, Bounded, Bounds, NumberConst, Udec128};
+
+/// Defines the bounds for a collateral power: 0 < CollateralPower <= 1.
+#[grug::derive(Serde)]
+pub struct CollateralPowerBounds;
+
+impl Bounds<Udec128> for CollateralPowerBounds {
+    const MAX: Option<Bound<Udec128>> = Some(Bound::Inclusive(Udec128::ONE));
+    const MIN: Option<Bound<Udec128>> = Some(Bound::Exclusive(Udec128::ZERO));
+}
+
+/// A decimal bounded by the collateral power bounds.
+pub type CollateralPower = Bounded<Udec128, CollateralPowerBounds>;
 
 /// The response type for a margin account's `Health` query.
 #[grug::derive(Serde)]
